move core functions to metrics folder
